// src/Data/sources/remote/models/ResponseMercadoPagoCardTocken.tsx
// Generated by https://quicktype.io

export interface ResponseMercadoPagoCardTocken{
    id:                     string;
    public_key:             string;
    first_six_digits:       string;
    expiration_month:       number;
    expiration_year:        number;
    last_four_digits:       string;
    cardholder:             Cardholder;
    status:                 string;
    date_created:           string;
    date_last_update:       string;
    date_due:               string;
    luhn_validation:        boolean;
    live_mode:              boolean;
    require_esc:            boolean;
    card_number_length:     number;
    security_code_length:   number;
}

export interface Cardholder {
    identification: Identification;
    name:           string;
}

export interface Identification {
    number: string;
    type:   string;
}