// src/Data/sources/remote/models/ResponseStripePayment.tsx
// Generated by https://quicktype.io

export interface ResponseStripePayment {
    id:              string;
    object:          string;
    billing_details: BillingDetails;
    card:            Card;
    created:         number;
    customer:        null;
    livemode:        boolean;
    type:            string;
}

export interface BillingDetails {
    address: Address;
    email:   null;
    name:    null;
    phone:   null;
}

export interface Address {
    city:        null;
    country:     null;
    line1:       null;
    line2:       null;
    postal_code: null;
    state:       null;
}

export interface Card {
    brand:                string;
    checks:               Checks;
    country:              string;
    exp_month:            number;
    exp_year:             number;
    funding:              string;
    generated_from:       null;
    last4:                string;
    networks:             Networks;
    three_d_secure_usage: ThreeDSecureUsage;
    wallet:               null;
}

export interface Checks {
    address_line1_check:       null;
    address_postal_code_check: null;
    cvc_check:                 null;
}

export interface Networks {
    available: string[];
    preferred: null;
}

export interface ThreeDSecureUsage {
    supported: boolean;
}